requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
httpx
requests
python-multipart
websockets